benchmark/os-harm/requirements.txt
numpy~=1.24.4
Pillow~=11.0.0
fabric
gymnasium~=0.28.1
requests~=2.31.0
transformers~=4.35.2
torch~=2.5.0
accelerate
opencv-python~=4.8.1.78
matplotlib~=3.7.4
pynput~=1.7.6
pyautogui~=0.9.54
psutil~=5.9.6
tqdm~=4.65.0
pandas~=2.2.3
flask~=3.0.0
requests-toolbelt~=1.0.0
lxml
cssselect
xmltodict
openpyxl
python-docx
python-pptx
pypdf
PyGetWindow
rapidfuzz
pyacoustid
pygame
opencv-python
ImageHash
scikit-image
librosa
pymupdf
chardet
playwright
backoff
formulas
pydrive
fastdtw
odfpy
openai
anthropic
func-timeout
beautifulsoup4
dashscope
google-generativeai
PyYaml
mutagen
easyocr
borb
pypdf2
pdfplumber
wandb
wrapt_timeout_decorator
gdown
tiktoken
groq
boto3
python-dotenv
azure-identity
azure-mgmt-compute
azure-mgmt-network
docker
loguru
